requirements.txt
numpy>=1.18.1
torch>=1.8.1
tqdm
PyYAML
click
